lift2/index.html

<html>
  <head>
    <link href="../tester/testharness.css" rel="stylesheet">
    <script src="../tester/testharness.js"></script>
  </head>
  <body>
    <style>
      body{
        margin: 0;
      }

      .wrapper{
        max-width: 1024px;
        margin: 0 auto;
        position: relative;
        height: 100%;
      }

      #lift{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        left: 40%;
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
        background-color: red;
      }

      #lift2{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        left: 60%;
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
        background-color: green;
      }

      #lift.floor0 {
        bottom: -150px;
      }
      #lift.floor1 {
        bottom: 0;
      }
      #lift.floor2 {
        bottom: 50%;
      }
      #lift.floor3 {
        bottom: calc(100% - 200px);
      }

      #lift2.floor20 {
        bottom: -150px;
      }
      #lift2.floor21 {
        bottom: 0;
      }
      #lift2.floor22 {
        bottom: 50%;
      }
      #lift2.floor23 {
        bottom: calc(100% - 200px);
      }


    </style>
    <div class='wrapper'>
      <div id='lift'></div>
      <div id='lift2'></div>
    </div>
    
    <p>first lift buttons</p>
    <div class='buttons buttons1'>
      <button class='floor3'>3</button>
      <button class='floor2'>2</button>
      <button class='floor1'>1</button>
      <button class='floor0'>-1</button>
    </div>
    
    <p>second lift buttons</p>
    <div class='buttons buttons2'>
      <button class='floor23'>3</button>
      <button class='floor22'>2</button>
      <button class='floor21'>1</button>
      <button class='floor20'>-1</button>
    </div>
    <br><br><br>
    <a class='call-lift' href='#'>call the lift</a>
  </body>

  <script>
    var lift = document.querySelector('#lift');
    var lift2 = document.querySelector('#lift2');
    var btn2 = document.querySelector('.floor2');
    var buttons = document.querySelectorAll('.buttons1 button');
    var buttons2 = document.querySelectorAll('.buttons2 button');
    var callLift = document.querySelector('.call-lift');

    lift.getFloor = function(){ return lift.className};

    btn2.addEventListener('click', function(e){
      lift.classList.add(e.target.className);
    });

    callLift.addEventListener('click', function(e){
      e.preventDefault();
      console.log(lift.className);
    });

    for (var i = 0; i < buttons.length; i++){
      buttons[i].addEventListener('click', function(e){
        lift.className = e.target.className;
      })
    }

    for (var i = 0; i < buttons2.length; i++){
      buttons2[i].addEventListener('click', function(e){
        lift2.className = e.target.className;
      })
    }

  </script>

<!--testing-->
  <script>

    test(function(){
      var lift = document.querySelector('#lift');
      assert_true(lift instanceof HTMLElement);
      assert_true(lift.getFloor instanceof Function);
    }, 'check if lift is exist');


    test(function(){
      var lift = document.querySelector('#lift');
      var floor = lift.getFloor();
      assert_equals(floor, lift.className);
    }, 'check where is the lift');

    var testBtn2Click = async_test('check if lift have class of clicked button');
    var lift = document.querySelector('#lift');
    var btn2 = document.querySelector('.floor2');

    btn2.addEventListener('click', testBtn2Click.step_func(function(){
      assert_true(lift.classList.contains('floor2'));
      this.done();
    }));

    btn2.dispatchEvent(new Event('click'));


    // var testBtnsClick = async_test('check if lift have classes of clicked buttons');
    var testBtn0Click = async_test('check if lift have classes of clicked button 0');
    var buttons = document.querySelectorAll('button');


    buttons[0].addEventListener('click', testBtn0Click.step_func(function(e){
      assert_true(lift.classList.contains(e.target.className));
      this.done();
    }));
    buttons[0].dispatchEvent(new Event('click'));


    var testBtn1Click = async_test('check if lift have classes of clicked button 1');
    buttons[1].addEventListener('click', testBtn1Click.step_func(function(e){
      assert_true(lift.classList.contains(e.target.className));
      this.done();
    }));
    buttons[1].dispatchEvent(new Event('click'));

    var testBtn2Click = async_test('check if lift have classes of clicked button 2');
    buttons[2].addEventListener('click', testBtn2Click.step_func(function(e){
      assert_true(lift.classList.contains(e.target.className));
      this.done();
    }));
    buttons[2].dispatchEvent(new Event('click'));


    var testBtn3Click = async_test('check if lift have classes of clicked button 3');
    buttons[3].addEventListener('click', testBtn3Click.step_func(function(e){
      assert_true(lift.classList.contains(e.target.className));
      this.done();
    }));
    buttons[3].dispatchEvent(new Event('click'));

    lift.className = '';


    var testCallLift = async_test('check if call button works');
    var lift = document.querySelector('#lift');
    var callLift = document.querySelector('.call-lift');
    var floor = lift.getFloor();

    callLift.addEventListener('click', testCallLift.step_func(function(e){
      e.preventDefault();
      assert_equals(floor, lift.className);
      this.done();
    }));
    callLift.dispatchEvent(new Event('click'));



  </script>
